chore: fix fullbuild bugs

+++ libc/src/sched/linux/sched_setcpuset.cpp
@@ -11,6 +11,7 @@
 #include "src/__support/common.h"        // LLVM_LIBC_FUNCTION
 #include "src/__support/macros/config.h" // LIBC_NAMESPACE_DECL
 
+#include "hdr/sched_macros.h" // NCPUBITS
 #include "hdr/types/cpu_set_t.h"
 #include "hdr/types/size_t.h"
 
@@ -18,9 +19,9 @@ namespace LIBC_NAMESPACE_DECL {
 
 LLVM_LIBC_FUNCTION(void, __sched_setcpuset,
                    (int cpu, const size_t cpuset_size, cpu_set_t *set)) {
-  if (cpu / 8 < cpuset_size) {
-    const size_t element_index = cpu / NCPUBITS;
-    const size_t bit_position = cpu % NCPUBITS;
+  if (static_cast<size_t>(cpu) / 8 < cpuset_size) {
+    const size_t element_index = static_cast<size_t>(cpu) / NCPUBITS;
+    const size_t bit_position = static_cast<size_t>(cpu) % NCPUBITS;
 
     const unsigned long mask = 1UL << bit_position;
     set->__bits[element_index] |= mask;
